* wip add table

* wip add edit/create form

* wip persons related to the project

* wip person form

* add important to form fields

* wip

* feat: add column on expense table

* add new style to table

* fix styles

* hotfix: update procedures author id

* wip

* hotfix: change styles on partner inputs

* hotfix: change styles on juridical partners inputs

* feat: register login and logout

* fix responsive on edit

* wip

* Feat task history (#90)

* wip: add custom history task changes log

* wip: add activity log on tasks

* feat: add task traceability with spatie activity log

* fix

* fix

* fix: activity log drop name migration class

* hotfix: login issue

* hotfix: login issue v2

* hotfix login v2

* wip fix revalidate

* feat: add not juridic person form

* remove console log

* fix store method

* fix pr issues

* fix partner validations

* fix: custom method for check relation change (#93)

* hotfix: only get update events on task changes view

* hotfix: add filter to index activity controller

* hotfix: add padding to partner inputs

* hotfix: add padding to partner inputs v2

* fix

* feat: add staffs section

* fix pr issues

* fix pr issue v2

* fix format on date

* Feat push notification (#95)

* wip: first commit add reminder push notification

* wip: add many device tokens to a staff

* feat: send notification from reminder if task is not closed

* feat: add reminders to procedure form and convert to task

* feat: add notification command to kernel and staff logged check

* fix

* feat: add store token on notification permission request

* feat: add dialog to accept web push notifications

* fix: notification api outside window validation

* feat: change default timezone from UTC to Panama

* fix: use notification permission variable instead of react state

* fix: create new task form on project tasks

* fix: procedures cache and sharp image optimization

* hotfix: flatpckr error on re-render and new reminder validation

* hotfix: process form edit author

* Fix errors (#97)

* fix: add label to show no data exist on table

* fix: see partners changes on proposal show

* fix: expense form date default value validation

* fix: date and type of task form variables

* fix: add effect to update the task timer

* fix: hide footer table if rows not exists

* fix: add label if task doesn't have changes

* fix: no store process cache on show

* fix: don't allow goes to next pagination page if doesn't exists

* Fix validation task form errors and invoice task tab form (#98)

* fix: add label to show no data exist on table

* fix: see partners changes on proposal show

* fix: expense form date default value validation

* fix: date and type of task form variables

* fix: add effect to update the task timer

* fix: hide footer table if rows not exists

* fix: invoice task tab form

* fix: add repeats fields validations on task form

* feat: add staff delete

* fix pr issues

* Feat process forks (#100)

* wip

* feat: add forks to process form (#96)

* feat: add front to pick the fork

* feat: add historic forks picks

* feat: filter for root forks

* feat: group tasks by process

* feat: add selected fork on show

* feat: send notification when a task is closed

* fix change colors of groups

* feat: send message if is the last process task and has been completed

* fix: check nulls on send notification when task is completed

* fix step number and milestone order on convertToTask method

* wip refactor content on show & add label to next step select

* fix: task level color group by process on project task tab

* fix selected processes view

* wip: select process on project task tab

* feat: add select to set next process fork

---------


* fix: hide select next process on project task tab

* fix revalidate path on staffs table

* fix show on tasks and edit on profile

* wip progress bar for processes

* feat: add striped color to project task tab table

* fix: change timezone to panama on min date restriction

* feat: add progress bar

* feat: add final node to selected process

* feat: add docassemble document generation (#103)

* feat: add docassemble document generation

* fix

* fix: total tasks size

* feat: add docassemble generate document in task action (#104)

* fix: invoice tasks tab view

* fix: bad redirection in email field on staff table

* fix: sale agent fk reference from partner to staff on estimates table

* fix: revalidate expenses view on task update

* fix: unify project loader

* fix: validation & select errors on partners form

* fix: remove open settings from dashboard navbar

* fix: form effect change when add file in a task

* fix: hide staff delete

---------

+++ back/app/Http/Controllers/TaskController.php
@@ -5,17 +5,25 @@ namespace App\Http\Controllers;
 use App\Http\Requests\TaskRequest;
 use App\Http\Resources\TaskResource;
 use App\Http\Resources\TaskResourceCollection;
+use App\Models\Staff;
 use App\Models\Taggable;
 use App\Models\Task;
 use App\Models\TaskStatus;
+use App\Pipes\TaskPipe;
+use App\Services\FcmService;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Spatie\Activitylog\Facades\CauserResolver;
 use Spatie\QueryBuilder\AllowedFilter;
 use Spatie\QueryBuilder\QueryBuilder;
 
 class TaskController extends Controller
 {
+    public function __construct(protected FcmService $fcmService)
+    {
+    }
+
     public function index()
     {
         $query = QueryBuilder::for(Task::class)
@@ -50,8 +58,7 @@ class TaskController extends Controller
                         'staff_id',
                         function (Builder $query, $value) {
                             $query
-                                ->where('owner_id', $value)
-                                ->orWhereHas(
+                                ->whereHas(
                                     'assigneds',
                                     fn (Builder $query) => $query->where('staff_id', $value)
                                 );
@@ -84,7 +91,7 @@ class TaskController extends Controller
         $newTask['author_id'] = $newTask['owner_id'];
 
         if (! array_key_exists('milestone_order', $newTask)) {
-            $newTask['milestone_order'] = Task::getMilestoneOrder($newTask['taskable_id'], $newTask['taskable_type']);
+            $newTask['milestone_order'] = Task::getLatestMilestoneOrder($newTask['taskable_id'], $newTask['taskable_type']) + 1; // Next milestone order
         }
 
         $task = Task::create($newTask);
@@ -114,41 +121,67 @@ class TaskController extends Controller
         $assigneds = isset($newTask['assigneds']) ? $newTask['assigneds'] : null;
         $followers = isset($newTask['followers']) ? $newTask['followers'] : null;
         $reminders = isset($newTask['reminders']) ? $newTask['reminders'] : null;
-        $actions = isset($newTask['actions']) ? $newTask['actions'] : null;
         $requiredFields = isset($newTask['requiredFields']) ? $newTask['requiredFields'] : null;
+
+        CauserResolver::setCauser(Staff::find($newTask['owner_id'] ?? $task->owner_id));
+
+        $task->load('tags', 'dependencies', 'comments', 'checklistItems', 'assigneds', 'followers', 'reminders', 'requiredFields');
+        $oldRelations = [
+            'tags' => $task->tags,
+            'dependencies' => $task->dependencies,
+            'comments' => $task->comments,
+            'checklistItems' => $task->checklistItems,
+            'assigneds' => $task->assigneds,
+            'followers' => $task->followers,
+            'reminders' => $task->reminders,
+            'requiredFields' => $task->requiredFields,
+        ];
+        $newRelations = [
+            'tags' => $tags,
+            'dependencies' => $dependencies,
+            'comments' => $comments,
+            'checklistItems' => $checklistItems,
+            'assigneds' => $assigneds,
+            'followers' => $followers,
+            'reminders' => $reminders,
+            'requiredFields' => $requiredFields,
+        ];
+
+        Task::addLogChange(new TaskPipe($oldRelations, $newRelations));
+
         $task->update($newTask);
 
-        if (isset($comments)) {
+        if ($comments) {
             $task->comments()->delete();
             $task->comments()->createMany($comments);
         }
 
-        if (isset($dependencies)) {
+        if ($dependencies) {
             $dependencyIds = array_column($dependencies, 'id');
             $task->dependencies()->sync($dependencyIds);
         }
 
-        if (isset($checklistItems)) {
+        if ($checklistItems) {
             $task->checklistItems()->delete();
             $task->checklistItems()->createMany($checklistItems);
         }
 
-        if (isset($assigneds)) {
+        if ($assigneds) {
             $assignedIds = array_column($assigneds, 'id');
             $task->assigneds()->sync($assignedIds);
         }
 
-        if (isset($followers)) {
+        if ($followers) {
             $followerIds = array_column($followers, 'id');
             $task->followers()->sync($followerIds);
         }
 
-        if (isset($reminders)) {
+        if ($reminders) {
             $task->reminders()->delete();
             $task->reminders()->createMany($reminders);
         }
 
-        if (isset($tags)) {
+        if ($tags) {
             $task->tags()->detach();
             foreach ($tags as $tag) {
                 $tag['taggable_id'] = $task->id;
@@ -158,13 +191,31 @@ class TaskController extends Controller
             }
         }
 
-        if (isset($actions)) {
-            $actionIds = array_column($actions, 'id');
-            $task->actions()->syncWithPivotValues($actionIds, ['is_completed' => false]);
+        if ($requiredFields) {
+            $task->requiredFields()->delete();
+            $task->requiredFields()->createMany($requiredFields);
         }
 
-        $task->requiredFields()->delete();
-        $task->requiredFields()->createMany($requiredFields);
+        if (isset($newTask['task_status_id']) && $newTask['task_status_id'] == TaskStatus::COMPLETED && $task->isFinalTask()) {
+            $staffs = Staff::whereIn('id',
+                array_merge(
+                    isset($newTask['assigneds']) ? $newTask['assigneds'] : array_column($task->assigneds->toArray(), 'id') ?? [],
+                    isset($newTask['followers']) ? $newTask['followers'] : array_column($task->followers->toArray(), 'id') ?? [],
+                    isset($newTask['owner_id']) ? [$newTask['owner_id']] : [$task->owner_id]
+                )
+            )->with('devices')->get();
+            foreach ($staffs as $staff) {
+                foreach ($staff->devices as $device) {
+                    $taskName = isset($newTask['name']) ? $newTask['name'] : $task->name;
+                    $this->fcmService->sendNotification(
+                        $device->device_token,
+                        'Tarea Completada',
+                        "La tarea \"$taskName\" ha sido completada, puede elegir el siguiente proceso",
+                        $staff->id
+                    );
+                }
+            }
+        }
 
         return response()->json(null, 204);
     }
@@ -213,6 +264,7 @@ class TaskController extends Controller
                 'actions',
                 'requiredFields',
                 'author',
+                'procedure',
             ])
             ->find($task->id);
